% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_survey_B}
\alias{df_survey_B}
\title{df_survey_B}
\format{
A data frame with 105242 rows and 29 variables.

\describe{
  \item{weight}{Survey weight}
  \item{psu}{Primary Sampling Unit}
  \item{ssu}{Secondary Sampling Unit}
  \item{strata}{Strata used for sampling}
  \item{ID}{ID number}
  \item{no_sample}{Sample number}
  \item{no_household}{Household number}
  \item{province}{Province code}
  \item{regency}{Regency or municipality code}
  \item{X1}{Predictor variables X1}
  \item{X2}{Predictor variables X2}
  \item{X3}{Predictor variables X3}
  \item{X4}{Predictor variables X4}
  \item{X5}{Predictor variables X5}
  \item{X6}{Predictor variables X6}
  \item{X7}{Predictor variables X7}
  \item{X8}{Predictor variables X8}
  \item{X9}{Predictor variables X9}
  \item{X10}{Predictor variables X10}
  \item{X11}{Predictor variables X11}
  \item{X12}{Predictor variables X12}
  \item{X13}{Predictor variables X13}
  \item{X14}{Predictor variables X14}
  \item{X15}{Predictor variables X15}
  \item{X16}{Predictor variables X16}
  \item{X17}{Predictor variables X17}
  \item{X18}{Predictor variables X18}
  \item{X19}{Predictor variables X19}
  \item{X20}{Predictor variables X20}
}
}
\usage{
df_survey_B
}
\description{
The dataset comes from a survey conducted at a more specific geographic level, corresponding to Domain 2.
}
\keyword{datasets}
