% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_svy1}
\alias{df_svy1}
\title{df_svy1: The August 2023 National Labor Force Survey dataset in Java Island, Indonesia.}
\format{
A data frame with 131.073 rows and 9 variables with 119 domains.

\describe{
  \item{id}{id}
  \item{prov}{province code}
  \item{regency}{regency/municipality code}
  \item{weight}{weight from survey}
  \item{strata}{strata}
  \item{edu}{last completed education}
  \item{age}{age}
  \item{sex}{sex (1: male, 2: female)}
  \item{head}{is she/he the head of the household}
}
}
\source{
\url{https://www.bps.go.id}
}
\usage{
df_svy1
}
\description{
A dataset containing several auxiliary variable in Java Island, Indonesia in 2023.
}
\keyword{datasets}
