% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_svy22}
\alias{df_svy22}
\title{df_svy22}
\format{
A data frame with 74.070 rows and 11 variables.

\describe{
  \item{PSU}{Primary Sampling Unit}
  \item{WEIGHT}{Weight from survey}
  \item{PROV}{province code}
  \item{REGENCY}{regency/municipality code}
  \item{STRATA}{Strata}
  \item{income}{Income}
  \item{neet}{Not in education employment or training status}
  \item{sex}{sex (1: male, 2: female)}
  \item{age}{age}
  \item{disability}{disability status (0: False, 1: True)}
  \item{edu}{last completed education}
}
}
\usage{
df_svy22
}
\description{
A dataset from a survey conducted at the province level in Indonesia in 2022.
}
\keyword{datasets}
