% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_svy_A}
\alias{df_svy_A}
\title{df_svy_A}
\format{
A data frame with 2000 rows and 20 variables with 40 domains.

\describe{
  \item{province}{Province code}
  \item{id_ind}{Unique identifier for each respondent}
  \item{num}{Sample number}
  \item{weight}{Weight from survey}
  \item{x1}{Predictor variables X1}
  \item{x2}{Predictor variables X2}
  \item{x3}{Predictor variables X3}
  \item{x4}{Predictor variables X4}
  \item{x5}{Predictor variables X5}
  \item{x6}{Predictor variables X6}
  \item{x7}{Predictor variables X7}
  \item{x8}{Predictor variables X8}
  \item{x9}{Predictor variables X9}
  \item{x10}{Predictor variables X10}
  \item{x11}{Predictor variables X11}
  \item{x12}{Predictor variables X12}
  \item{x13}{Predictor variables X13}
  \item{x14}{Predictor variables X14}
  \item{x15}{Predictor variables X15}
  \item{Y}{Target variable (1: Yes, 0: No)}
}
}
\usage{
df_svy_A
}
\description{
A simulation dataset from a small sample survey, presented only at provincial level (Domain 1).
}
\keyword{datasets}
