% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_estimator.R
\name{projection_estimator}
\alias{projection_estimator}
\title{Projection Estimator}
\usage{
projection_estimator(
  formula,
  domain,
  weight,
  data_model,
  data_proj,
  model,
  cutoff = "optimum",
  ...
)
}
\arguments{
\item{formula}{An object of class formula that contains a description of the model to be fitted. The variables included in the formula must be contained in the \code{data_model} dan \code{data_proj}.}

\item{domain}{Column names in data_model and data_proj representing specific domains for which disaggregated data needs to be produced.}

\item{weight}{Column name in data_proj representing the survey weight.}

\item{data_model}{A data frame or a data frame extension (e.g., a tibble) representing the second survey, characterized by a much smaller sample, provides information on both the variable of interest and the auxiliary variables.}

\item{data_proj}{A data frame or a data frame extension (e.g., a tibble) representing the first survey, characterized by a large sample that collects only auxiliary information or general-purpose variables.}

\item{model}{The working model to be used in the projection estimator. The available working models are "linear" for linear regression and "logistic" for logistic regression.}

\item{cutoff}{Cutoff for binary classification (model='logistic')}

\item{...}{Further argument to the \code{\link[stats]{lm}} and \code{\link[stats]{glm}} model.}
}
\value{
The function returns a list with the following objects (\code{model}, \code{prediction} and \code{df_result}):
\code{model} model yang digunakan dalam projection
\code{prediction} A vector containing the prediction results from the model.
\code{df_result} A data frame with the following columns:
\itemize{
\item \code{domain} The name of the domain.
\item \code{yhat} The estimated result for each domain.
\item \code{var_yhat} The sample variance of the projection estimator for each domain.
\item \code{rse} The Relative Standard Error (RSE) in percentage (\%).
}
}
\description{
The function addresses the problem of combining information from two or more independent surveys, a common challenge in survey sampling. It focuses on cases where: \cr
\itemize{
\item Survey 1: A large sample collects only auxiliary information.
\item Survey 2: A much smaller sample collects data on both the variables of interest and the auxiliary variables.
}
The function implements a model-assisted projection estimation method based on a working model, with the reference distribution derived from a design-based approach.
}
\examples{
\dontrun{
library(sae.projection)
library(dplyr)

df_res <- projection_estimator(
   formula = unemployment ~ edu + sex + age + head,
   domain = c('prov', 'regency'),
   weight = 'weight',
   data_model = df_svy2,
   data_proj = df_svy1,
   model = 'logistic'
)

df_svy2_new <- df_svy2 \%>\% dplyr::filter(!is.na(income))

m2 <- projection_estimator(
  formula = income ~ edu + sex + age + head,
  domain = c('prov', 'regency'),
  weight = 'weight',
  data_model = df_svy2_new,
  data_proj = df_svy1,
  model = 'linear'
)
}
}
\references{
\enumerate{
\item Kim, J. K., & Rao, J. N. (2012). Combining data from two independent surveys: a model-assisted approach. Biometrika, 99(1), 85-100.
}
}
