% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_svy_A}
\alias{df_svy_A}
\title{df_svy_A}
\format{
A data frame with 3655 rows and 33 variables with 6 domains.

\describe{
  \item{ID}{Unique identifier for each respondent}
  \item{no_sample}{Sample number}
  \item{no_household}{Household number}
  \item{no_member}{Household member number}
  \item{weight}{Weight from survey}
  \item{province}{Province code}
  \item{X1}{Predictor variables X1}
  \item{X2}{Predictor variables X2}
  \item{X3}{Predictor variables X3}
  \item{X4}{Predictor variables X4}
  \item{X5}{Predictor variables X5}
  \item{X6}{Predictor variables X6}
  \item{X7}{Predictor variables X7}
  \item{X8}{Predictor variables X8}
  \item{X9}{Predictor variables X9}
  \item{X10}{Predictor variables X10}
  \item{X11}{Predictor variables X11}
  \item{X12}{Predictor variables X12}
  \item{X13}{Predictor variables X13}
  \item{X14}{Predictor variables X14}
  \item{X15}{Predictor variables X15}
  \item{X16}{Predictor variables X16}
  \item{X17}{Predictor variables X17}
  \item{X18}{Predictor variables X18}
  \item{X19}{Predictor variables X19}
  \item{X20}{Predictor variables X20}
  \item{X21}{Predictor variables X21}
  \item{X22}{Predictor variables X22}
  \item{X23}{Predictor variables X23}
  \item{X24}{Predictor variables X24}
  \item{X25}{Predictor variables X25}
  \item{X26}{Predictor variables X26}
  \item{Y}{Target variable (1: Yes, 0: No)}
}
}
\usage{
df_svy_A
}
\description{
A dataset from a survey that was conducted in a certain province within Indonesia, presented only at provincial level (Domain 1).
}
\keyword{datasets}
